Fix pmap tracer leak, merge gibbs_util into gibbs, type-check touched modules
- Fix TracerArrayConversionError in DiscreteGibbs.init and MixedHMC.init under
  pmap: ravel_pytree stages a jnp.concatenate even on constant inputs inside
  pmap's staging trace, so np.asarray of its result raised on CI where the
  upgraded JAX exposes multiple CPU devices by default. The flat support sizes
  are now built with pure numpy (_flat_support_sizes, ravel_pytree leaf order),
  keeping them static on the kernel. Adds a device-independent regression test
  (single-chain pmap over init/sample for DiscreteGibbs, MixedHMC,
  DiscreteHMCGibbs) and runs test_gibbs.py -k chain under the forced
  multi-device CI step. Note: the stored dtype changes int32 -> int64 (benign,
  consumers re-wrap with jnp.asarray).
- Merge numpyro/infer/gibbs_util.py into numpyro/infer/gibbs.py (helpers first,
  discrete proposals next to DiscreteGibbs); update importers and doc refs.
- Add _typing.py, infer/gibbs.py, infer/hmc.py, infer/hmc_gibbs.py,
  infer/mcmc.py, infer/mixed_hmc.py to the ty include list and fix all 47
  diagnostics: widen numpyro.util.cond's pred annotation to bool | jax.Array,
  drop the None sentinel of Gibbs._sites in favor of (), assert-narrow lazily
  initialized attributes, rename HMC.postprocess_fn params to match the base
  class, and swap mixed_hmc's deprecated 5-arg lax.cond call for numpyro's own
  cond wrapper (also honors NUMPYRO_DISABLE_CONTROL_FLOW_PRIM). make lint now
  type-checks every module this PR touches.
